move stuff to the correct dirs
