+++ packages/lib/src/types/MemoWithGenerics.types.ts
@@ -1,10 +1,26 @@
-/**
- * Constructs a type which provides generic typing for React.memo
- */
-
 import { ReactElement } from 'react';
 
-// This removes propTypes and displayName
+/**
+ * Type signature that enhances React.memo to support generics more effectively.
+ *
+ * This removes the public properties `propTypes` and `displayName`.
+ *
+ * You can still use `propTypes` but you have to silence the TS error with `@ts-expect-error`.
+ *
+ * `displayName` can be replaced by a named function.
+ *
+ * E.g.
+ *
+ * ```
+ * const memoWithGenerics: MemoWithGenerics = React.memo;
+ *
+ * export const MemoizedRow = memoWithGenerics(Row, (prevProps, nextProps) => {})
+ *
+ * // @ts-expect-error
+ * MemoizedRow.propTypes = {};
+ * ```
+ */
+
 export interface MemoWithGenerics {
   <P extends object>(
     Component: (props: P) => ReactElement | null,
