+++ packages/table/src/utils/genericHelpers.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { ForwardRefWithGenerics, MemoWithGenerics } from '@leafygreen-ui/lib';
+
+export const forwardRefWithGenerics: ForwardRefWithGenerics = React.forwardRef;
+
+export const memoWithGenerics: MemoWithGenerics = React.memo;
